requirements.txt
langchain
langchain-community
pypdf
chromadb
openai
gradio
tiktoken
python-dotenv
